+++ src/components/Header/index.tsx
@@ -5,6 +5,7 @@ import { CommonHeaderProps } from './common';
 import { TrainLCDAppState } from '../../store';
 import { AppTheme } from '../../store/types/theme';
 import HeaderYamanote from '../HeaderYamanote';
+import HeaderJRWest from '../HeaderJRWest';
 
 const Header = ({
   station,
@@ -32,6 +33,18 @@ const Header = ({
     case AppTheme.Yamanote:
       return (
         <HeaderYamanote
+          state={state}
+          station={station}
+          stations={stations}
+          nextStation={nextStation}
+          line={line}
+          lineDirection={lineDirection}
+          boundStation={boundStation}
+        />
+      );
+    case AppTheme.JRWest:
+      return (
+        <HeaderJRWest
           state={state}
           station={station}
           stations={stations}
