+++ src/store/reducers/theme.ts
@@ -5,7 +5,7 @@ export interface ThemeState {
 }
 
 const initialState: ThemeState = {
-  theme: AppTheme.Yamanote, // FIXME: TokyoMetroに戻す
+  theme: AppTheme.JRWest, // FIXME: TokyoMetroに戻す
 };
 
 const themeReducer = (
