+++ src/store/types/theme.ts
@@ -3,6 +3,7 @@ export const UPDATE_THEME = 'UPDATE_THEME';
 export enum AppTheme {
   TokyoMetro,
   Yamanote,
+  JRWest,
 }
 
 interface UpdateThemeActionPayload {
